Add polling-based project log code (#204)
Fixes #188 (in an MVP form)

This PR covers the REST version of fetching docker logs, which is the easy
one that comes with built in "starting at what datetime?" functionality.
Hooking websockets into the docker log ouput so it updates if and when
there's more log data is going to need more work, and should not hold
up landing this, because really any sort of logs insight is better than no
logs insight.

+++ src/server/docker/docker-helpers.js
@@ -1,5 +1,6 @@
 import { sep } from "node:path";
 import {
+  BYPASS_DOCKER,
   CONTENT_BASE,
   getFreePort,
   STARTER_BASE,
@@ -21,6 +22,8 @@ import { scheduleScreenShot } from "../screenshots/screenshot.js";
  * ...docs go here...
  */
 export function checkContainerHealth(project, slug = project.slug) {
+  if (BYPASS_DOCKER) return `not running`;
+
   const check = `docker ps --no-trunc -f name=^/${slug}$`;
   const result = execSync(check).toString().trim();
   if (result.includes(`Exited`)) {
@@ -42,6 +45,8 @@ export function checkContainerHealth(project, slug = project.slug) {
  * ...docs go here...
  */
 export function deleteContainer(project, slug = project.slug) {
+  if (BYPASS_DOCKER) return;
+
   try {
     execSync(`docker container rm ${slug}`, { stdio: `ignore` });
   } catch (e) {
@@ -58,6 +63,8 @@ export function deleteContainer(project, slug = project.slug) {
  * ...docs go here...
  */
 export function deleteContainerAndImage(project) {
+  if (BYPASS_DOCKER) return;
+
   console.log(`removing container and image...`);
   stopContainer(project);
   deleteContainer(project);
@@ -67,6 +74,8 @@ export function deleteContainerAndImage(project) {
  * ...docs go here...
  */
 export function getAllRunningContainers() {
+  if (BYPASS_DOCKER) return [];
+
   const containerData = [];
   const output = execSync(`docker ps -a --no-trunc --format json`)
     .toString()
@@ -99,9 +108,32 @@ export function getAllRunningStaticServers() {
 }
 
 /**
+ * Get a container's log ouput, optionally "since some specific time"
+ */
+export function getContainerLogs(project, since = 0) {
+  if (BYPASS_DOCKER) return false;
+  const { slug } = project;
+  const cmd = `docker container logs --since ${since} ${slug}`;
+  try {
+    const output = execSync(cmd).toString();
+    // And now for the stupid part:
+    const datetime = execSync(cmd.replace(`--since`, `-t --since`))
+      .toString()
+      .split(`\n`)
+      .at(-2)
+      .replace(/Z.*/, `Z`);
+    return { output, datetime };
+  } catch {
+    return false;
+  }
+}
+
+/**
  * ...docs go here...
  */
 export function renameContainer(oldSlug, newSlug) {
+  if (BYPASS_DOCKER) return;
+
   stopContainer(oldSlug);
   try {
     execSync(`docker tag ${oldSlug} ${newSlug}`);
@@ -113,6 +145,8 @@ export function renameContainer(oldSlug, newSlug) {
  * ...docs go here...
  */
 export async function restartContainer(project, rebuild = false) {
+  if (BYPASS_DOCKER) return;
+
   const { slug } = project;
   if (rebuild) {
     console.log(`rebuilding container for ${slug}...`);
@@ -122,7 +156,6 @@ export async function restartContainer(project, rebuild = false) {
     console.log(`restarting container for ${slug}...`);
     try {
       execSync(`docker container restart -t 0 ${slug}`);
-      portBindings[slug].restarts ??= 0;
       portBindings[slug].restarts++;
     } catch (e) {
       // if an admin force-stops this container, we can't "restart".
@@ -136,6 +169,13 @@ export async function restartContainer(project, rebuild = false) {
  * ...docs go here...
  */
 export async function runContainer(project, slug = project.slug) {
+  if (BYPASS_DOCKER) {
+    if (project.settings?.app_type === `static`) {
+      runStaticServer(project);
+    }
+    return;
+  }
+
   // note: we assume the caller already checked for project
   // suspension, so we don't try to use the database here.
   const isStarter = isStarterProject(project);
@@ -232,15 +272,21 @@ export async function runStaticServer(project) {
   ].join(` `);
   const runCommand = `node src/server/static.js ${opts}`;
   if (TESTING) console.log({ runCommand });
-  const child = exec(runCommand, { shell: true, stdio: `inherit` });
   const binding = updateCaddyFile(project, port);
-  binding.serverProcess = child;
+  binding.serverProcess = exec(runCommand);
 }
 
 /**
  * ...docs go here...
  */
 export function stopContainer(project, slug = project.slug) {
+  if (BYPASS_DOCKER) {
+    if (project.settings?.app_type === `static`) {
+      stopStaticServer(project);
+    }
+    return;
+  }
+
   try {
     execSync(`docker container stop ${slug}`, { stdio: `ignore` });
   } catch (e) {
